make it an external hook
